Functions w/o parameters.
Functions are represented by CompiledFunction(offset) values where
offset is the entry point of the function in the bytecode.

Implemented CALL, RETURN, and PUSHFN instructions.

PUSHFN pushes the compiled function object into the stack which is
consumed by CALL.

CALL and RETURN modify the ip and set the current frame as required.